crawler/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
openai==1.51.0
supabase==2.7.4
python-dotenv==1.0.1
Pillow==10.4.0
